@@ -1,5 +1,7 @@
 package waits;
 
+import java.time.Duration;
+import java.util.List;
 import java.util.stream.Collectors;
 
 import org.openqa.selenium.By;
@@ -9,7 +11,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 // import org.openqa.selenium.firefox.FirefoxDriver;
 // import org.openqa.selenium.edge.EdgeDriver;
 
-public class threadSleep {
+public class implicitWait {
 
     public static void wait(int millisegundos) {
         try {
@@ -36,12 +38,16 @@ public class threadSleep {
 		// WebDriver driver = new EdgeDriver();
     	
 
+
+        // Define o tempo de espera implicito, se elemento nao for encontrado no tempo especificado um erro é levantado
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
+
+
+
         String baseUrl = "https://ubuntu.com/";
 
         driver.get(baseUrl);
         
-        wait(5000);
-
         WebElement botaoAceitar = driver.findElements(By.tagName("button"))
                                         .stream()
                                         .filter(botao -> botao.getText().equals("Accept all and visit site"))
@@ -51,10 +57,7 @@ public class threadSleep {
         // WebElement botaoAceitar = driver.findElement(By.xpath("//button[text() = 'Accept all and visit site']"));
 
         botaoAceitar.click();       
-        
-        wait(5000);
-        
-
+               
         WebElement dropDownDownload = driver.findElements(By.className("p-navigation__link-anchor"))
                                         .stream()
                                         .filter(botao -> botao.getText().equals("Download"))
@@ -65,19 +68,26 @@ public class threadSleep {
 
         dropDownDownload.click();    
 
-        wait(5000);
-
-        WebElement desktopUbuntu = driver.findElements(By.tagName("a"))
+        List<WebElement> desktopUbuntu = driver.findElements(By.tagName("a"))
                                         .stream()
                                         .filter(link -> link.getText().startsWith("Ubuntu Desktop"))
-                                        .collect(Collectors.toList())
-                                        .get(0);
+                                        .collect(Collectors.toList());
+
+        int nroTentativas = 1;
+        while (desktopUbuntu.isEmpty() || (!desktopUbuntu.isEmpty() && !desktopUbuntu.get(0).isDisplayed() && !desktopUbuntu.get(0).isEnabled())){
+            desktopUbuntu = driver.findElements(By.tagName("a"))
+                                .stream()
+                                .filter(link -> link.getText().startsWith("Ubuntu Desktop"))
+                                .collect(Collectors.toList());
+            nroTentativas +=1;
+        }
+
+        System.out.println(String.format("Foi preciso processar %d vezes para poder clicar no link", nroTentativas));
+
 
         // WebElement desktopUbuntu = driver.findElement(By.xpath("//a[contains(text(), 'Ubuntu Desktop')]/parent::h4"));
         
-        desktopUbuntu.click(); 
-
-        wait(5000);
+        desktopUbuntu.get(0).click(); 
 
         WebElement botaoDownload = driver.findElement(By.className("is-wide"));
 
@@ -85,8 +95,8 @@ public class threadSleep {
 
         botaoDownload.click();
         
-        wait(5000);
 
+        wait(5000);
         driver.close();
     }
 
